+++ arms-20190808/alibabacloud_arms20190808/client.py
@@ -133,6 +133,17 @@ class Client(OpenApiClient):
         return self.add_grafana_with_options(request, runtime)
 
     def add_integration_with_options(self, request, runtime):
+        """
+        @deprecated : AddIntegration is deprecated, please use ARMS::2019-08-08::InstallAddon instead.
+        
+
+        @param request: AddIntegrationRequest
+
+        @param runtime: runtime options for this request RuntimeOptions
+
+        @return: AddIntegrationResponse
+        Deprecated
+        """
         UtilClient.validate_model(request)
         query = {}
         if not UtilClient.is_unset(request.cluster_id):
@@ -161,6 +172,15 @@ class Client(OpenApiClient):
         )
 
     def add_integration(self, request):
+        """
+        @deprecated : AddIntegration is deprecated, please use ARMS::2019-08-08::InstallAddon instead.
+        
+
+        @param request: AddIntegrationRequest
+
+        @return: AddIntegrationResponse
+        Deprecated
+        """
         runtime = util_models.RuntimeOptions()
         return self.add_integration_with_options(request, runtime)
 
@@ -1146,6 +1166,9 @@ class Client(OpenApiClient):
 
     def create_integration_with_options(self, request, runtime):
         UtilClient.validate_model(request)
+        query = {}
+        if not UtilClient.is_unset(request.region_id):
+            query['RegionId'] = request.region_id
         body = {}
         if not UtilClient.is_unset(request.auto_recover):
             body['AutoRecover'] = request.auto_recover
@@ -1158,6 +1181,7 @@ class Client(OpenApiClient):
         if not UtilClient.is_unset(request.recover_time):
             body['RecoverTime'] = request.recover_time
         req = open_api_models.OpenApiRequest(
+            query=OpenApiUtilClient.query(query),
             body=OpenApiUtilClient.parse_to_map(body)
         )
         params = open_api_models.Params(
@@ -2537,6 +2561,17 @@ class Client(OpenApiClient):
         return self.delete_imrobot_with_options(request, runtime)
 
     def delete_integration_with_options(self, request, runtime):
+        """
+        @deprecated : DeleteIntegration is deprecated, please use ARMS::2019-08-08::DeleteAddonRelease instead.
+        
+
+        @param request: DeleteIntegrationRequest
+
+        @param runtime: runtime options for this request RuntimeOptions
+
+        @return: DeleteIntegrationResponse
+        Deprecated
+        """
         UtilClient.validate_model(request)
         query = {}
         if not UtilClient.is_unset(request.cluster_id):
@@ -2565,6 +2600,15 @@ class Client(OpenApiClient):
         )
 
     def delete_integration(self, request):
+        """
+        @deprecated : DeleteIntegration is deprecated, please use ARMS::2019-08-08::DeleteAddonRelease instead.
+        
+
+        @param request: DeleteIntegrationRequest
+
+        @return: DeleteIntegrationResponse
+        Deprecated
+        """
         runtime = util_models.RuntimeOptions()
         return self.delete_integration_with_options(request, runtime)
 
@@ -3821,6 +3865,17 @@ class Client(OpenApiClient):
         return self.get_grafana_workspace_with_options(request, runtime)
 
     def get_integration_state_with_options(self, request, runtime):
+        """
+        @deprecated : GetIntegrationState is deprecated, please use ARMS::2019-08-08::DescribeAddonRelease instead.
+        
+
+        @param request: GetIntegrationStateRequest
+
+        @param runtime: runtime options for this request RuntimeOptions
+
+        @return: GetIntegrationStateResponse
+        Deprecated
+        """
         UtilClient.validate_model(request)
         query = {}
         if not UtilClient.is_unset(request.cluster_id):
@@ -3849,6 +3904,15 @@ class Client(OpenApiClient):
         )
 
     def get_integration_state(self, request):
+        """
+        @deprecated : GetIntegrationState is deprecated, please use ARMS::2019-08-08::DescribeAddonRelease instead.
+        
+
+        @param request: GetIntegrationStateRequest
+
+        @return: GetIntegrationStateResponse
+        Deprecated
+        """
         runtime = util_models.RuntimeOptions()
         return self.get_integration_state_with_options(request, runtime)
 
@@ -5413,6 +5477,8 @@ class Client(OpenApiClient):
         query = {}
         if not UtilClient.is_unset(request.addon_name):
             query['AddonName'] = request.addon_name
+        if not UtilClient.is_unset(request.bind_resource_id):
+            query['BindResourceId'] = request.bind_resource_id
         if not UtilClient.is_unset(request.environment_type):
             query['EnvironmentType'] = request.environment_type
         if not UtilClient.is_unset(request.region_id):