+++ arms-20190808/ChangeLog.md
@@ -1,3 +1,9 @@
+2024-01-30 Version: 4.2.1
+- Update API CreateIntegrationadd RegionId param.
+- Update API DescribeEnvironmentupdate response param.
+- Update API ListEnvironmentsadd BindResourceId param.
+
+
 2024-01-26 Version: 4.2.0
 - Generated python2 2019-08-08 for ARMS.